+++ test/Setup.t.sol
@@ -360,7 +360,7 @@ contract Setup is Test {
             switchBoardAddress_,
             DEFAULT_BATCH_LENGTH,
             uint32(remoteChainSlug_),
-            uint32(capacitorType_)
+            capacitorType_
         );
 
         scc_.siblingChainSlug = remoteChainSlug_;
